chassis_controller/pyproject.toml
[tool.poetry]
name = "avant_hardware_interface_fastapi"
version = "0.1.0"
description = ""
authors = ["Aaron Kehrer <akehrer@in2being.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
pyserial = "^3.5"
crcmod = "^1.7"
aioserial = "^1.3.1"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
